requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
numpy==1.26.2
pillow==10.1.0
scikit-learn==1.3.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
matplotlib==3.8.2

